+++ lib/src/core/resources/world_objects/actors/i_actor_data.dart
@@ -11,6 +11,8 @@ import 'package:dawnforge/src/core/systems/drop/drop_entry.dart';
 class IActorData extends IWorldObjectData {
   IActorData({
     required super.id,
+    super.displayNameKey,
+    super.descriptionKey,
     super.spritesheetPath,
     super.frameWidth,
     super.frameHeight,
@@ -96,6 +98,8 @@ class IActorData extends IWorldObjectData {
   @override
   IActorData clone() => IActorData(
         id: id,
+        displayNameKey: displayNameKey,
+        descriptionKey: descriptionKey,
         spritesheetPath: spritesheetPath,
         frameWidth: frameWidth,
         frameHeight: frameHeight,